+++ cypress-tests/integration-tests/pages/InteractiveReportPage.ts
@@ -5,14 +5,11 @@ export default class InteractiveReportPage {
 
   updateFiltersDetails = (): PageElement => cy.xpath(`//*[@id="dpr-interactive-filters-details"]`)
 
-  selectedFilterOne = (): PageElement =>
-    cy.xpath(`//*[@id="main-content"]/div/div/div[1]/div/div[5]/div[2]/div[1]/div/a[1]`)
+  selectedFilterOne = (): PageElement => cy.xpath(`//*[@id="dpr-selected-filters"]/a[1]`)
 
-  selectedFilterTwo = (): PageElement =>
-    cy.xpath(`//*[@id="main-content"]/div/div/div[1]/div/div[3]/div[2]/div[1]/div/a[2]`)
+  selectedFilterTwo = (): PageElement => cy.xpath(`//*[@id="dpr-selected-filters"]/a[2]`)
 
-  selectedFilterThree = (): PageElement =>
-    cy.xpath(`//*[@id="main-content"]/div/div/div[1]/div/div[5]/div[2]/div[1]/div/a[3]`)
+  selectedFilterThree = (): PageElement => cy.xpath(`//*[@id="dpr-selected-filters"]/a[3]`)
 
   applyFiltersButton = () => cy.xpath(`//*[@id="interactive-apply-filters-button"]`)
