+++ src/dpr/components/_filters/filters-selected/style.scss
@@ -6,6 +6,7 @@
     padding: govuk-spacing(4);
 
     .interactive-remove-filter-button {
+      cursor: pointer;
       background-color: #fff;
       border: 1px solid #0b0c0c;
       color: #0b0c0c;
